requirements.txt
pillow==12.0.0
PySocks==1.7.1
pyTelegramBotAPI==4.29.1
requests==2.32.5
telebot==0.0.5
